Rename tree.h into hodlr.h
