+++ src/hodlr_algebra.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "../include/tree.h"
+#include "../include/hodlr.h"
 #include "../include/error.h"
 #include "../include/utils.h"
 #include "../include/blas_wrapper.h"
