+++ src/utils.c
@@ -1,4 +1,4 @@
-#include "../include/tree.h"
+#include "../include/hodlr.h"
 #include "../include/utils.h"
